add delete dialog

+++ src/app/backend/manage-books/manage-books.component.ts
@@ -1,7 +1,9 @@
 import { SelectionModel } from '@angular/cdk/collections';
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
+import { DeleteDialogComponent } from 'src/app/dialog/delete-dialog/delete-dialog.component';
 import { CommonService } from 'src/app/services/common.service';
 import { SnackBarService } from 'src/app/services/snackbar.service';
 
@@ -16,12 +18,15 @@ export class ManageBooksComponent implements OnInit {
   hasData: boolean = false;
   userdata: any = [];
   selection = new SelectionModel<any>(true, []);
+  @Output() closeDialog = new EventEmitter<void>();
 
-  displayedColumns = ['isbn', 'title', 'publication', 'price', 'eddition', 'action'];
+  displayedColumns = ['id', 'isbn', 'title', 'publication', 'price', 'eddition', 'action'];
 
   constructor(private commonService: CommonService,
               private router: Router,
-              private snackBar: SnackBarService) { }
+              private snackBar: SnackBarService,
+              public dialog: MatDialog
+            ) { }
 
   ngOnInit(): void {
     this.loaddata();
@@ -46,23 +51,33 @@ export class ManageBooksComponent implements OnInit {
       action: 'delete',
     };
   }
-
   delete(id: string) {
-    const payload = this.deleteUser(id);
-    this.commonService.delete_data_operation(payload).subscribe(
-      (response) => {
-        if (response.status === 'success') {
-          let message = response.message;
-          this.snackBar.openSnackBarSuccess([message]);
-          this.loaddata();
-        } else {
-          this.snackBar.openSnackBarError([response.message]);
-        }
-      },
-      (error) => {
-        console.error(error);
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      width: '350px',
+      height: '200px',
+      data: { id: id }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        const payload = this.deleteUser(id);
+        this.commonService.delete_data_operation(payload).subscribe(
+          (response) => {
+            if (response.status === 'success') {
+              let message = response.message;
+              this.snackBar.openSnackBarSuccess([message]);
+              this.loaddata();
+              this.closeDialog.emit();
+            } else {
+              this.snackBar.openSnackBarError([response.message]);
+            }
+          },
+          (error) => {
+            console.error(error);
+          }
+        );
       }
-    );
+    });
   }
 
   applyFilter(event: Event) {
